requirements.txt
flask>=3.1.0
flask-socketio>=5.5.0
pillow>=12.0.0
twitchio>=3.1.0
python-socketio>=5.11.0
pyyaml>=6.0
requests>=2.31.0
